+++ tests/objects/ready/test_ready.py
@@ -6,4 +6,4 @@ def test_simple():
     data = samples.example_connect
 
     obj = Ready()
-    obj.ingest_raw_dict(data)
+    obj.from_dict(data)