+++ docs/changelog.md
@@ -2,7 +2,7 @@
 
 This project follows [Semantic Versioning](https://semver.org/) and [Keep A Change Log](https://keepachangelog.com/en/1.0.0/).
 
-## [v0.4.0] WIP
+## [v0.4.0] 2021-10-27
 ### Added
 - New set_all_intents() function in DiscordClient. Sets all intents to True.
 - Better example code.